+++ Code/VFSPrototype/VFSBase/VFSBase.csproj
@@ -56,6 +56,8 @@
     <Compile Include="Implementation\BlockAllocation.cs" />
     <Compile Include="Persistence\Coding\EncryptionOptions.cs" />
     <Compile Include="Persistence\Coding\NullStreamCodingStrategy.cs" />
+    <Compile Include="Persistence\Coding\SelfMadeCaesarCryptor.cs" />
+    <Compile Include="Persistence\Coding\SelfMadeSimpleCryptor.cs" />
     <Compile Include="Persistence\Coding\SelfMadeStreamEncryptionStrategy.cs" />
     <Compile Include="Persistence\Coding\StreamCompressionEncryptionCodingStrategy.cs" />
     <Compile Include="Persistence\Coding\MicrosoftStreamCompressionStrategy.cs" />
